Search autoscaling: update executor CPU quantity for load calculation (#2226)
This PR updates the CPU quantity to be fractional for search load calculations.

+++ x-pack/plugin/stateless/src/main/java/org/elasticsearch/xpack/stateless/autoscaling/search/load/ExecutorLoadStats.java
@@ -22,5 +22,5 @@ public record ExecutorLoadStats(
     double averageTaskExecutionEWMA,
     int currentQueueSize,
     int maxThreads,
-    int numProcessors
+    double numProcessors
 ) {}
